requirements.txt
PyMuPDF>=1.24.0
Pillow>=10.2.0
pytesseract>=0.3.10
deep-translator>=1.11.4
